Delete a large amount of unnecessary code. Most of it was not used, with the main exception being the use of the ExtractionPlant in the IDW profile. The IDW's use of ExtractionPlant involved only the beneficiation stage, and the logic from there has been moved into an IDW implementation of the IMetadataExtractor interface.
