--- src/beneficiations/beneficiation_consts.py
@@ -1,5 +1,0 @@
-"""Constants used in beneficiation.
-"""
-
-JSON_FORMAT = "json"
-YAML_FORMAT = "yaml"
