+++ src/profiles/idw/profile.py
@@ -5,11 +5,9 @@ Ingestion profile for ingestions using the Instrument Data Wizard (IDW)
 from pathlib import Path
 from typing import Optional
 
-from src.beneficiations.beneficiation import Beneficiation
-from src.extraction.extraction_plant import ExtractionPlant
 from src.extraction.metadata_extractor import IMetadataExtractor
 from src.profiles.idw.cleanup import idw_cleanup
-from src.profiles.idw.custom_beneficiation import CustomBeneficiation  # type: ignore
+from src.profiles.idw.metadata_extraction import IDWMetadataExtractor
 from src.profiles.profile_base import IProfile
 
 
@@ -24,11 +22,7 @@ class IDWProfile(IProfile):
         return "idw"
 
     def get_extractor(self) -> IMetadataExtractor:
-        return ExtractionPlant(
-            prospector=None,
-            miner=None,
-            beneficiation=Beneficiation(beneficiation=CustomBeneficiation()),
-        )
+        return IDWMetadataExtractor()
 
     def cleanup(self, source_data_path: Path) -> None:
         return idw_cleanup(source_data_path)
