+++ src/profiles/idw/yaml_helper.py
@@ -8,10 +8,10 @@ appropriate dataclasses.
 """
 
 import logging
-import os
 
 # Standard library imports
 from pathlib import Path, PosixPath
+from typing import Any, Dict, List, Union
 
 # Third-party imports
 import yaml
@@ -23,7 +23,6 @@ from yaml.nodes import MappingNode, Node
 from src.blueprints import RawDatafile, RawDataset, RawExperiment, RawProject
 from src.blueprints.common_models import GroupACL, UserACL
 from src.blueprints.custom_data_types import Username
-from src.miners.utils import datafile_metadata_helpers
 from src.mytardis_client.common_types import DataClassification, DataStatus
 
 # Constants
@@ -46,6 +45,48 @@ tags = [
     username_tag,
     path_tag,
 ]
+
+
+def replace_micrometer_values(
+    data: Union[Dict[Any, Any], List[Any]], replacement: str
+) -> Union[Dict[Any, Any], List[Any]]:
+    """Recursively replace micrometer values in a dictionary or list.
+
+    This function searches for string values in dictionaries or lists that end with "µm"
+    (micrometers) and replaces them with the specified replacement string.
+
+    Args:
+        data (Union[Dict, List]): The dictionary or list containing values to check and replace.
+        replacement (str): The replacement string for micrometer values.
+
+    Returns:
+        Union[Dict, List]: The input data with micrometer values replaced.
+    """
+    if isinstance(data, dict):
+        for key, value in data.items():
+            data[key] = (
+                replace_micrometer_values(value, replacement)
+                if isinstance(value, (dict, list))
+                else (
+                    value[:-2] + replacement
+                    if isinstance(value, str) and value.endswith("µm")
+                    else value
+                )
+            )
+    elif isinstance(data, list):
+        return [
+            (
+                replace_micrometer_values(item, replacement)
+                if isinstance(item, (dict, list))
+                else (
+                    item[:-2] + replacement
+                    if isinstance(item, str) and item.endswith("µm")
+                    else item
+                )
+            )
+            for item in data
+        ]
+    return data
 
 
 class YamlParser:
@@ -204,9 +245,7 @@ class YamlParser:
         # data["md5sum"] = md5sum
         metadata = data.pop("metadata", {})
         # TO DO: change back when debugging from the API side
-        metadata_replaced = datafile_metadata_helpers.replace_micrometer_values(
-            metadata, "um"
-        )
+        metadata_replaced = replace_micrometer_values(metadata, "um")
         data["metadata"] = metadata_replaced
         datafile = RawDatafile(**data)
         return datafile
