--- src/prospectors/__init__.py
@@ -1,1 +1,0 @@
-# pylint: disable=missing-module-docstring
