+++ tests/test_biru_beneficiation.py
@@ -1,7 +1,7 @@
 import logging
 from pathlib import Path
 
-from src.profiles.idw.custom_beneficiation import CustomBeneficiation
+from src.profiles.idw.metadata_extraction import IDWMetadataExtractor
 
 logger = logging.getLogger(__name__)
 logger.propagate = True
@@ -9,9 +9,9 @@ logger.propagate = True
 
 def test_beneficiate() -> None:
     """Tests that beneficiation for BIRU is successful and extracts metadata."""
-    a = CustomBeneficiation()
-    fpath = "tests/testdata/ingestion.yaml"
-    ingestible_dataclasses = a.beneficiate(Path(fpath))
+    extractor = IDWMetadataExtractor()
+    fpath = Path("tests/testdata/ingestion.yaml")
+    ingestible_dataclasses = extractor.extract(fpath)
     df = ingestible_dataclasses.get_datafiles()[0]
     assert df.filename == "20221113_slide3-2_humanRWM_cd34_x20_0.12umpix_8.czi"
     assert df.metadata and df.metadata["Experimenter|UserName"] == "hsuz002"
